# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Auditor IA API
  version: 1.0.0
  description: API local para auditorias persistidas de negócios do Bitrix24.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, example: AUDIT_NOT_FOUND}
            message: {type: string, example: Auditoria não encontrada}
    Pagination:
      type: object
      required: [page, page_size, total, total_pages]
      properties:
        page: {type: integer, example: 1}
        page_size: {type: integer, example: 20}
        total: {type: integer, example: 1}
        total_pages: {type: integer, example: 1}
    Audit:
      type: object
      required: [assessment_id, bitrix_deal_id, deal_title, status, score, final_result, result_source, summary, findings_count, started_at, created_at]
      properties:
        assessment_id: {type: integer, format: int64, example: 42}
        bitrix_deal_id: {type: integer, format: int64, example: 8620}
        deal_title: {type: string, example: Lead 8620}
        status: {type: string, enum: [PENDING, PROCESSING, COMPLETED, FAILED]}
        score: {type: number, nullable: true, minimum: 0, maximum: 100, example: 80}
        final_result: {type: string, example: venda}
        result_source: {type: string, example: estágio ganho no Bitrix}
        summary: {type: string}
        findings_count: {type: integer, example: 2}
        started_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time, nullable: true}
        created_at: {type: string, format: date-time}
        trace_id: {type: string, pattern: '^[0-9a-f]{32}$', nullable: true}
        conversation_status: {type: string, enum: [NO_CONVERSATION, EMPTY_CONVERSATION, ACCESS_DENIED, AVAILABLE]}
        conversation_message: {type: string, description: Motivo seguro quando a análise conversacional está indisponível.}
    AuditEnvelope:
      type: object
      properties:
        data: {$ref: '#/components/schemas/Audit'}
    AuditList:
      type: object
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/Audit'}
        pagination: {$ref: '#/components/schemas/Pagination'}
    Finding:
      type: object
      properties:
        id: {type: integer, format: int64}
        assessment_id: {type: integer, format: int64}
        bitrix_deal_id: {type: integer, format: int64}
        rule: {type: string, example: opportunity_without_value}
        description: {type: string}
        entity_type: {type: string, example: deal}
        entity_id: {type: string, example: '8620'}
        severity: {type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL]}
        created_at: {type: string, format: date-time}
    FindingList:
      type: object
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/Finding'}
        pagination: {$ref: '#/components/schemas/Pagination'}
...
  responses:
    BadRequest:
      description: Requisição inválida
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unauthorized:
      description: API key inválida ou ausente
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
    Internal:
      description: Erro interno
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    External:
      description: Falha no Bitrix ou Ollama
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unavailable:
      description: Serviço temporariamente indisponível
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
security:
  - ApiKeyAuth: []
paths:
  /api/v1/audit-batches:
    post:
      summary: Cria auditoria em lote persistente
      security: [{ApiKeyAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bitrix_deal_ids]
              properties:
                bitrix_deal_ids:
                  type: array
                  minItems: 1
                  maxItems: 10000
                  uniqueItems: true
                  items: {type: integer, format: int64, minimum: 1}
      responses:
        '202': {description: Lote persistido e aguardando workers}
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/audits:
    post:
      summary: Executa uma auditoria síncrona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bitrix_deal_id]
              properties: {bitrix_deal_id: {type: integer, format: int64, minimum: 1, example: 8620}}
      responses:
        '201': {description: Auditoria concluída, content: {application/json: {schema: {$ref: '#/components/schemas/AuditEnvelope'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/Internal'}
        '502': {$ref: '#/components/responses/External'}
        '503': {$ref: '#/components/responses/Unavailable'}
    get:
      summary: Lista o histórico de avaliações
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: bitrix_deal_id, in: query, schema: {type: integer, format: int64}}
        - {name: status, in: query, schema: {type: string, enum: [PENDING, PROCESSING, COMPLETED, FAILED]}}
        - {name: final_result, in: query, schema: {type: string}}
        - {name: date_from, in: query, schema: {type: string, format: date}}
        - {name: date_to, in: query, schema: {type: string, format: date}}
      responses:
        '200': {description: Histórico paginado, content: {application/json: {schema: {$ref: '#/components/schemas/AuditList'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/Internal'}
...
  /api/v1/findings:
    get:
      summary: Lista achados persistidos
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: assessment_id, in: query, schema: {type: integer, format: int64}}
        - {name: bitrix_deal_id, in: query, schema: {type: integer, format: int64}}
        - {name: severity, in: query, schema: {type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL]}}
        - {name: rule, in: query, schema: {type: string}}
      responses:
        '200': {description: Achados paginados, content: {application/json: {schema: {$ref: '#/components/schemas/FindingList'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/Internal'}
...